refactor: move render template into skills, separate skills and grammar loading
